Implement trace ID ratio-based sampling for OpenTelemetry
